third_party/WebKit/LayoutTests/http/tests/devtools/service-workers/service-worker-v8-cache.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/timeline-test.js"></script>
<script src="../../inspector/service-workers/service-workers-test.js"></script>
<script>

function loadScript() {
  const url = 'v8-cache-script.js';
  const frameId = 'frame_id';
  let iframeWindow = document.getElementById(frameId).contentWindow;
  return iframeWindow.loadScript(url)
    .then(() => iframeWindow.loadScript(url));
}

function test() {
  const scriptURL = 'resources/v8-cache-worker.js';
  const scope = 'resources/v8-cache-iframe.html';
  const frameId = 'frame_id';

  ApplicationTestRunner.registerServiceWorker(scriptURL, scope)
      .then(_ => ApplicationTestRunner.waitForActivated(scope))
      .then(() => {
        return TestRunner.addIframe(scope, {id: frameId});
      })
      .then(() => {
        // Need to suspend targets, because V8 doesn't produce the cache when
        // the debugger is loaded.
        return SDK.targetManager.suspendAllTargets();
      })
      .then(() => {
        return new Promise(
            (r) => {
              PerformanceTestRunner.invokeAsyncWithTimeline('loadScript', r);
            });
      })
      .then(() => {
        PerformanceTestRunner.printTimelineRecordsWithDetails(
            TimelineModel.TimelineModel.RecordType.CompileScript);
        TestRunner.completeTest();
      });
}

</script>
</head>
<body onload="runTest()">
<p>Tests V8 cache information of Service Worker Cache Storage in timeline</p>
</body>
</html>
