third_party/WebKit/LayoutTests/http/tests/devtools/service-workers/service-worker-v8-cache-expected.txt

Tests V8 cache information of Service Worker Cache Storage in timeline


v8.compile Properties:
{
    data : {
        cacheProduceOptions : "code"
        columnNumber : <number>
        lineNumber : <number>
        producedCacheSize : <number>
        streamed : <boolean>
        url : .../devtools/service-workers/resources/v8-cache-script.js
    }
    endTime : <number>
    startTime : <number>
    type : "v8.compile"
}
Text details for v8.compile: v8-cache-script.js:1
v8.compile Properties:
{
    data : {
        cacheConsumeOptions : "code"
        cacheRejected : false
        columnNumber : <number>
        consumedCacheSize : <number>
        lineNumber : <number>
        streamed : <boolean>
        url : .../devtools/service-workers/resources/v8-cache-script.js
    }
    endTime : <number>
    startTime : <number>
    type : "v8.compile"
}
Text details for v8.compile: v8-cache-script.js:1
